refactor: extract Markdown parsing functions into proper files
