+++ src/types/markdown.ts
@@ -1,0 +1,12 @@
+export interface Frontmatter {
+  id: string;
+  title: string;
+  subtitle: string;
+  tags: string[];
+  stack: string[];
+}
+
+export interface MarkdownFile {
+  frontmatter: Frontmatter;
+  content: string;
+}
